Basic Features Implemented, without any interconnection
main.py: extracts the images from the camera using live view dump image and a viewport ('lvdumpimg -vp') at 640x360 in RGB as .ppm files. If you instead use ('remoteshoot -jpg') or ('shoot -dl -rm') you will result in 4000x3000 in RGB as .jpg files, however it takes a long time to take continuous photos (approx. 1-3 seconds in between each shutter).

HandTrackingMin.py: uses opencvs hand tracking solution to track one's hands. (Currently not implemented with the video feed that comes out of the camera)

imagetovideo.py: uses opencv to convert ppm file to a frame, which is then collected and converted into a video file (for testing purposes and is out of the scope of the project)